Organized the file structures to make it cleaner and easier to understand
